+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -15,6 +15,8 @@
   to CompositeTextMapPropagator. IPropagator is renamed to TextMapPropagator
   and changed from interface to abstract class.
   ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1427))
+* Added GlobalPropagators API via Propagators.DefaultTextMapPropagator.
+  ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1428))
 
 ## 0.7.0-beta.1